requirements.txt
openai~=0.28.0
flask~=3.0.2
py_eureka_client
python-dotenv~=1.0.1
requests~=2.31.0
